Moved search results display to own file
